Added things to some levels and solved problems with the Vive
- Added many little things
- Now the elevator rides support the Vive, meaning that the camera will
go up with the elevator; instead of using the CVmCamera object now it
only uses glTranslate(), so that the effects will also appear when using
the Vive
- I found out another problem with the Vive, that involved the starting
position of the player. Basically the player can move only in a 4x4
square in real life, and in the game this means that also the camera
sticks to those values. This means that if we want the player to start
in a certain position, we have to move every objects in the scene so
that it "looks like" he's in that place; so in my levels I changed every
object's position using a fixxed translate factor, which is added to
every object's position that need it